refactor: slice.js에서 사용한 모듈 분리
공통 사용이 많은 전역 함수를 util.js로 이동
