+++ src/slice.js
@@ -1,24 +1,13 @@
 import { createSlice } from "@reduxjs/toolkit";
 
-const LATEX_LIST = "latexList";
+import {
+	LATEX_LIST,
+	getLocalStorage,
+	updateSidebar,
+	addLatexItem,
+} from "./util";
 
-const saveLocalStorage = (key, value) => localStorage.setItem(key, JSON.stringify(value));
-const getLocalStorage = key => JSON.parse(localStorage.getItem(key)) || [];
-const getIdToAdd = list => list.reduce((maxId, { id }) => (maxId < id ? id : maxId), 0) + 1;
-const updateSidebar = state => {
-	state.latexList = state.latexList.filter(item => item.isRecent || item.isBookmark);
-	state.bookmarkItems = state.latexList.filter(item => item.isBookmark);
-	state.recentItems = state.latexList.filter(item => item.isRecent);
-	saveLocalStorage(LATEX_LIST, state.latexList);
-};
-const addLatexItem = (state, { latex, isRecent = false, isBookmark = false }) => {
-	const id = getIdToAdd(state.latexList);
-	const newItem = { id, latex, isRecent, isBookmark };
-
-	state.latexList.push(newItem);
-};
-
-const latexList = getLocalStorage(LATEX_LIST);
+const latexList = getLocalStorage(LATEX_LIST, []);
 
 const { reducer, actions } = createSlice({
 	name: "FEditor",
